pyproject.toml
[tool.poetry]
name = "whist"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
pytest = "^7.1.3"
ipython = "^8.5.0"
pytest-django = "^4.5.2"
vue = "^0.0.1"
node = "^1.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
pytest-xdist = "^3.0.2"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
